+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 * Search invoice by PO number. Refs UINV-162.
 * Add hover text to invoice link icon. Refs UINV-290.
 * Increment stripes to v7. Refs UINV-284.
+* Invoices - Implement MCL Next/Previous pagination. UINV-293.
 
 ## [2.4.4]https://github.com/folio-org/ui-invoice/tree/v2.4.4) (2021-09-08)
 [Full Changelog](https://github.com/folio-org/ui-invoice/compare/v2.4.3...v2.4.4)